Implement View Rendering [finished #113227033]
Why?
The controller actions should be able to call a method `render` to render a view they wish to display and pass some parameters to it.

How?
Create the `controller.rb` file in the dependencies and define the `Controller` class in it. define the render method in this class and the consuming application's controllers inherit from this controller

https://www.pivotaltracker.com/story/show/113227033