Implement Request Object in Controller [finished #113232967]
Why?
An application should be able to access the request object in the controller

How?
pass the `env` parameter to the Controller upon initializing it
Use `Rack::Request.new(env)` in the controller to build the request object
define `params` method in the controller to return the params attribute of the request object

https://www.pivotaltracker.com/story/show/113232967