Implement Route Handling [finished #113088395]
Why?
A developer should be able to create application routes/endpoints for their project by calling the `get`, `post`, `put`, `patch` and `delete` in a block to the `create` method of the `Router` class
Also allow developers to create resource routes by calling the `resources` method

How?
Create the `Router` class that has the verb methods
Create the `create` method in the `Router` class that takes a block and evaluates it in the instance of this class using `instance_eval`
Create the `Routes` class that builds an instance of a route with the attributes of a given route
Create the `RequestHandler` method that responds to a request and process it
Create the `RouteHelper` module and include it in the `Router` class. The module provides the extra methods for routing like the `root` routing and the `resources` routing
Use these methods in the sampe app `Next Watch`

https://www.pivotaltracker.com/story/show/113088395