Write Tests [finished #113700203]
Why?
To make sure the code works as its expected, there is need to write test to asert what we expect from a particular section of the code and the wholistic usage in general

How?
Install `rspec` and `capybara` gems
Write test with the DSL provided by the said gems

https://www.pivotaltracker.com/story/show/113700203